feat: consolidate collection schema types into @mulmoclaude/collection-plugin (chunk 1b)
Move the canonical collection schema types (CollectionSchema, CollectionFieldSpec/
Type, CollectionAction, CollectionWhen, CollectionCustomView, CollectionSpawn,
CollectionDetail, CollectionItem, …) from server/workspace/collections/types.ts
into the package core as the single source of truth, shared by the host server,
the host frontend, and MulmoTerminal.

- server `collections/types.ts` is now a re-export barrel from the package.
- frontend `collectionTypes.ts` re-exports the canonical types under the names it
  has always used (FieldType ← CollectionFieldType, FieldSpec ← CollectionFieldSpec)
  and keeps its UI-only view-state types (TableRowDraft/EditState/RefCache/…).
  The frontend types were a structural subset, so all 13 importers compile unchanged.

Feeds decoupling: the canonical CollectionSchema types `ingest` as the minimal
CollectionIngest (kind/url/schedule) so the package stays independent of the host
feeds subsystem. The feeds-only IngestSpec now `extends CollectionIngest`; the
engine narrows feed schemas via a small `feedIngest()` helper (feeds are validated
to carry the rich spec at discovery).

Verified: yarn typecheck (server/vue/test/packages), lint, build all green;
384 collection + feeds unit tests pass.
