+++ packages/plugins/collection-plugin/src/index.ts
@@ -5,6 +5,7 @@
 // MulmoClaude and MulmoTerminal share one implementation and the server no
 // longer reaches into `src/` for it. Vue surfaces will live in `./vue`.
 
+export * from "./core/schema";
 export * from "./core/actionVisible";
 export * from "./core/derivedFormula";
 export * from "./core/deriveAll";
