+++ src/components/collectionTypes.ts
@@ -1,148 +1,29 @@
-// Shared types for the collection surfaces — the list/detail view
-// (CollectionView.vue), the extracted record panel
-// (CollectionRecordPanel.vue), the calendar view
-// (CollectionCalendarView.vue), and the rendering composable
-// (composables/collections/useCollectionRendering.ts). These lived
-// inline in CollectionView.vue until the calendar view needed the panel
-// to be reusable; the shapes are unchanged.
+// Collection surface types for the list/detail view (CollectionView.vue), the
+// record panel (CollectionRecordPanel.vue), the calendar view, and the
+// rendering composable (composables/collections/useCollectionRendering.ts).
+//
+// The canonical schema types (field types, field/action specs, schema, custom
+// views, detail) now live in @mulmoclaude/collection-plugin — the single source
+// of truth shared with the server and MulmoTerminal. They are re-exported here
+// under the names the frontend has always used (FieldType ← CollectionFieldType,
+// FieldSpec ← CollectionFieldSpec) so the 13 importing components keep compiling
+// unchanged. This file additionally owns the UI-only view-state types, which
+// have no server/storage analog.
 
-export type FieldType =
-  | "string"
-  | "text"
-  | "email"
-  | "number"
-  | "date"
-  | "datetime"
-  | "boolean"
-  | "markdown"
-  | "ref"
-  | "money"
-  | "enum"
-  | "table"
-  | "derived"
-  | "image"
-  | "file"
-  | "embed"
-  | "toggle";
+import type { CollectionDetail, CollectionItem, CollectionSchema } from "@mulmoclaude/collection-plugin";
 
-export interface FieldSpec {
-  type: FieldType;
-  label: string;
-  primary?: boolean;
-  required?: boolean;
-  /** When type === "ref" or "embed": slug of the target collection. */
-  to?: string;
-  /** When type === "embed": primary-key value of the fixed record
-   *  pulled from `to` and rendered read-only in the detail view. */
-  id?: string;
-  /** When type === "money" (or derived/money): a literal ISO 4217
-   *  currency, fixed for every record. */
-  currency?: string;
-  /** When type === "money" (or derived/money): name of a sibling
-   *  record field holding the ISO code, so currency can vary per record. */
-  currencyField?: string;
-  /** When type === "enum": closed list of allowed string values. */
-  values?: readonly string[];
-  /** When type === "table": sub-schema for each row. */
-  of?: Record<string, FieldSpec>;
-  /** When type === "derived": formula evaluated against the record. */
-  formula?: string;
-  /** When type === "derived": render the computed value as this type. */
-  display?: FieldType;
-  /** Optional visibility predicate: render this field only when
-   *  `String(record[when.field])` is one of `when.in`. */
-  when?: { field: string; in: string[] };
-  /** When type === "toggle": the `enum` field this checkbox projects
-   *  (stores nothing itself — reads + writes that field). */
-  field?: string;
-  /** When type === "toggle": the enum value meaning "checked" (and written
-   *  on check). */
-  onValue?: string;
-  /** When type === "toggle": the enum value written on uncheck. */
-  offValue?: string;
-}
+export type {
+  CollectionFieldType as FieldType,
+  CollectionFieldSpec as FieldSpec,
+  CollectionAction,
+  CollectionViewCapability,
+  CollectionCustomView,
+  CollectionSchema,
+  CollectionDetail,
+  CollectionItem,
+} from "@mulmoclaude/collection-plugin";
 
-/** A schema-declared, per-record action rendered as a button in the
- *  detail view. */
-export interface CollectionAction {
-  id: string;
-  label: string;
-  icon?: string;
-  kind: "chat";
-  role: string;
-  template: string;
-  when?: { field: string; in: string[] };
-}
-
-/** What a custom view's capability token may do against the data endpoint.
- *  Mirror of `CollectionViewCapability` in server collections types. */
-export type CollectionViewCapability = "read" | "write";
-
-/** A custom (LLM-authored) HTML view. Mirror of `CollectionCustomView` in
- *  `server/workspace/collections/types.ts`. */
-export interface CollectionCustomView {
-  id: string;
-  label: string;
-  icon?: string;
-  file: string;
-  capabilities?: CollectionViewCapability[];
-}
-
-export interface CollectionSchema {
-  title: string;
-  icon: string;
-  dataPath: string;
-  primaryKey: string;
-  /** When set, the collection is a singleton: at most one record whose
-   *  primary key is fixed to this value. */
-  singleton?: string;
-  fields: Record<string, FieldSpec>;
-  actions?: CollectionAction[];
-  /** Collection-level actions rendered as buttons in the header. Carry no
-   *  record context; the `when` predicate is not evaluated. */
-  collectionActions?: CollectionAction[];
-  /** Name of the field whose value labels the record in notifications and
-   *  the calendar chip (falls back to the primary key). */
-  displayField?: string;
-  /** Name of a `date` field anchoring the optional calendar view. When
-   *  unset, the toggle still appears if any `date` field exists. */
-  calendarField?: string;
-  /** Name of a second `date` field marking the END of a multi-day span
-   *  on the calendar. Requires `calendarField`. */
-  calendarEndField?: string;
-  /** Name of a string field holding a free-form time or time-range
-   *  (e.g. "14:00-17:00", "17:00-", "16:30") used to place records on the
-   *  day (time-allocation) view. Only consulted when the calendar date
-   *  fields are date-only (a `datetime` anchor/end pair carries its own
-   *  clock and takes precedence). Requires `calendarField`. */
-  calendarTimeField?: string;
-  /** Name of an `enum` field grouping records into columns on the optional
-   *  Kanban board. When unset, the toggle still appears if any `enum` field
-   *  exists (the first one, in declaration order, is the default and is
-   *  switchable in-view). */
-  kanbanField?: string;
-  /** Optional predicate gating the completion bell (server-side); reuses
-   *  the `when` shape. */
-  notifyWhen?: { field: string; in: string[] };
-  /** Optional custom (LLM-authored) HTML views, each rendered in a
-   *  sandboxed iframe over the records. */
-  views?: CollectionCustomView[];
-  /** Present only on "feed" collections (the <workspace>/feeds/ registry):
-   *  declarative retrieval config the host uses to refill the records.
-   *  When set, the view shows a Refresh control. */
-  ingest?: { kind: string; url: string; schedule: string };
-}
-
-export interface CollectionDetail {
-  slug: string;
-  title: string;
-  icon: string;
-  // "feed" = a data-source collection from the <workspace>/feeds/ registry.
-  source: "user" | "project" | "feed";
-  schema: CollectionSchema;
-}
-
-export type CollectionItem = Record<string, unknown>;
+// ── UI-only types (frontend view state; no server/storage analog) ──
 
 /** A record file the server couldn't load or that violates the schema —
  *  silently skipped at read time. Mirror of `RecordIssue` in
